factor out top-selectors
